Add `role` property to User model and database schema
Updated the `User.cs` file to include a new `role` property with a maximum length of 50 characters. The property is mapped to the `role` column in the database.

Modified the `01-create-users-table.sql` file to add a `role` column to the `users` table. The column is defined as `VARCHAR(50)` with a default value of `'guest'`.

These changes enable the application to manage user roles while maintaining backward compatibility.